# api/swagger/v1/redeam-rest.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Book service",
    "version": "1.0",
    "contact": {
      "name": "redeam-rest",
      "url": "https://github.com/radean0909/redeam-rest",
      "email": "radean0909@gmail.com"
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/book/all": {
      "get": {
        "operationId": "ReadAll",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ReadAllResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "api",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "BookService"
        ]
      }
    },
...
  },
  "definitions": {
    "BookStatus": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "CHECKED_IN",
        "CHECKED_OUT"
      ],
      "default": "UNKNOWN"
    },
    "v1Book": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "title": {
          "type": "string"
        },
        "author": {
          "type": "string"
        },
        "publisher": {
          "type": "string"
        },
        "publish_date": {
          "type": "string",
          "format": "date-time"
        },
        "rating": {
          "type": "number",
          "format": "double"
        },
        "status": {
          "$ref": "#/definitions/BookStatus"
        }
      },
      "title": "Books the library has"
    },
...
    "v1ReadAllResponse": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string"
        },
        "books": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Book"
          }
        }
      }
    },
...
  }
}